Added methods:
append(int)  - The append method should append the parameter to the end
of the list. If the list was full, then increase the size by 50% so
there will be room. Increment the count.

first(): int  - Return the first element in the list

size(): int  - Return the current number of possible locations in the
list